+++ .storybook/preview.jsx
@@ -1,9 +1,10 @@
-import React from 'react'
-import { ThemeProvider } from '../src/providers'
-import componentTheme from '../example/theme'
-
-import breakpoints from '../styles/_breakpoints.module.scss'
 import './preview.scss'
+
+import React from 'react'
+
+import componentTheme from '../example/theme'
+import { ThemeProvider } from '../src/providers'
+import breakpoints from '../styles/_breakpoints.module.scss'
 
 const viewports = [Number(breakpoints.desktopLarge)]
 
@@ -17,12 +18,7 @@ export const decorators = [
 
 export const parameters = {
   chromatic: { viewports },
-  darkMode: {
-    current: 'dark'
-  },
-  options: {
-    storySort: {
-      order: ['Docs', ['README', 'CHANGELOG'], 'Components']
-    }
-  }
+  options: { storySort: { order: ['Docs', ['README', 'CHANGELOG']] } }
 }
+
+export const tags = ['autodocs']