+++ .storybook/theme.js
@@ -1,7 +1,6 @@
-import { create } from '@storybook/theming/create'
+import { create } from 'storybook/theming/create'
 
 export default create({
-  base: 'dark',
   brandTitle: 'Chipset',
   brandUrl: '/',
   brandImage: '/logo.svg'