+++ src/__mocks__/fixtures.js
@@ -1,5 +1,0 @@
-const baseUrl = '/fixtures/'
-
-export const imageUrl = `${baseUrl}image.jpg`
-export const imageWebPUrl = `${baseUrl}image.webp`
-export const videoUrl = `${baseUrl}video.mp4`