+++ src/components/Element/Element.jsx
@@ -3,9 +3,6 @@ import React from 'react'
 
 const Element = ({ as: Component = 'div', ...rest }) => <Component {...rest} />
 
-Element.propTypes = {
-  as: oneOfType([func, string]),
-  children: node
-}
+Element.propTypes = { as: oneOfType([func, string]), children: node }
 
 export default Element