+++ src/components/Image/Image.stories.jsx
@@ -1,16 +1,14 @@
 import { imageUrl, imageWebPUrl } from '../../__mocks__/fixtures'
 import { Image } from '..'
 
-export default {
-  component: Image
-}
+export default { component: Image }
 
 export const Source = {
   args: {
-    src: 'https://Example.com/image.jpg',
+    src: 'https://example.com/image.jpg',
     sources: [
-      { srcSet: 'https://Example.com/image.webp' },
-      { srcSet: 'https://Example.com/image.jpg' }
+      { srcSet: 'https://example.com/image.webp' },
+      { srcSet: 'https://example.com/image.jpg' }
     ]
   },
   parameters: { chromatic: { disable: true } }