+++ src/components/SmartLink/SmartLink.spec.jsx
@@ -13,7 +13,13 @@ describe('Components/SmartLink', () => {
   it("should render a <button /> when no 'href' is set", () => {
     render(<SmartLink>Content</SmartLink>)
 
-    expect(screen.getByRole('button')).toBeInTheDocument()
+    expect(screen.getByRole('button')).toBeEnabled()
+  })
+
+  it("should render a disabled <button /> when no 'href' and 'disabled' is set", () => {
+    render(<SmartLink disabled>Content</SmartLink>)
+
+    expect(screen.getByRole('button')).toBeDisabled()
   })
 
   it("should render an <a /> when internal 'href' is set", () => {
@@ -30,7 +36,7 @@ describe('Components/SmartLink', () => {
   })
 
   it("should render an <a /> when external 'href' is set", () => {
-    render(<SmartLink href="https://test.com/">Content</SmartLink>)
+    render(<SmartLink href="https://example.com/">Content</SmartLink>)
 
     const link = screen.getByRole('link')
 
@@ -40,7 +46,7 @@ describe('Components/SmartLink', () => {
 
   it("should add 'rel=noopener noreferrer' when external 'href' and 'target=blank' is set", () => {
     render(
-      <SmartLink href="https://test.com/" target="_blank">
+      <SmartLink href="https://example.com/" target="_blank">
         Content
       </SmartLink>
     )
@@ -49,5 +55,18 @@ describe('Components/SmartLink', () => {
 
     expect(link).toBeInTheDocument()
     expect(link).toHaveAttribute('rel', 'noopener noreferrer')
+  })
+
+  it("should render a disabled <a /> when 'disabled' is set", () => {
+    render(
+      <SmartLink href="/" disabled>
+        Content
+      </SmartLink>
+    )
+
+    const link = screen.getByRole('link')
+
+    expect(link).toBeInTheDocument()
+    expect(link).toHaveAttribute('aria-disabled', 'true')
   })
 })