+++ src/components/VisuallyHidden/VisuallyHidden.stories.jsx
@@ -1,8 +1,10 @@
 import { VisuallyHidden } from '..'
 
-export default {
-  component: VisuallyHidden,
+export default { component: VisuallyHidden }
+
+export const Source = {
+  args: { children: 'Content' },
   parameters: { chromatic: { disable: true } }
 }
 
-export const Source = { args: { children: 'Content' } }
+export const Example = { args: Source.args }